Refactors templates to pass directory
Also switched to throwing exception when template does not exist.
This should allow for a better interface (fail before hitting Mailgun
instead of sending along a bad message).
It also allows the template directory to be configured at runtime, which
will work better with an uberjar and adding new templates without
restarting.